create default utility file
