+++ utils/index.js
@@ -1,0 +1,17 @@
+import { DEFAULT_WARE_IMAGE } from './constants'
+
+export const configure_services = ({ data, path }) => {
+  return data?.map(ware => {
+    const img = ware.promo_image
+      ? { src: ware.promo_image, alt: `The promotional image for ${ware.name}` }
+      : DEFAULT_WARE_IMAGE
+
+    return {
+      description: ware.snippet,
+      id: ware.reference_of_id,
+      img,
+      name: ware.name,
+      href: `${path}/${ware.slug}`,
+    }
+  })
+}
